requirements.txt
fastapi
uvicorn[standard]
jinja2

asyncpg
python-dotenv
slowapi
httpx
sendgrid
pydantic[email]